portfolio-agent/requirements.txt
python-dotenv>=1.0.1
python-docx>=1.1.2
pypdf>=5.1.0
pytest>=8.3.3
